OCT-111: moved SearchAttributesQuery
