+++ components/catalogs/back/src/Infrastructure/Controller/Internal/GetAttributesAction.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Infrastructure\Controller\Internal;
 
-use Akeneo\Catalogs\Application\Persistence\SearchAttributesQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Attribute\SearchAttributesQueryInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
